+++ libs/elodin-editor/src/ui/images.rs
@@ -24,6 +24,7 @@ pub struct Images {
     pub icon_exit_fullscreen: Handle<Image>,
     pub icon_setting: Handle<Image>,
     pub icon_lightning: Handle<Image>,
+    pub icon_link: Handle<Image>,
 }
 
 impl FromWorld for Images {
@@ -58,6 +59,7 @@ impl FromWorld for Images {
             icon_setting: asset_server.load("embedded://elodin_editor/assets/icons/setting.png"),
             icon_lightning: asset_server
                 .load("embedded://elodin_editor/assets/icons/lightning.png"),
+            icon_link: asset_server.load("embedded://elodin_editor/assets/icons/link.png"),
         }
     }
 }